requirements.txt
python-resize-image~=1.1.19
Pillow~=8.0.1